feat: implement search strategy factory and add grid search strategy
