--- src/backend/automl/search/__init__.py
@@ -1,13 +1,0 @@
-from .base import SearchStrategy
-from .factory import SearchStrategyFactory
-from .strategies.grid_search import GridSearchStrategy
-
-# Đăng ký strategy mặc định
-SearchStrategyFactory.register_strategy('grid', GridSearchStrategy)
-# SearchStrategyFactory.register_strategy('random', RandomSearchStrategy)
-
-# Auto-discover thêm các strategy trong thư mục strategies
-SearchStrategyFactory.discover_strategies()
-
-# Export các thành phần chính
-__all__ = ['SearchStrategy', 'SearchStrategyFactory', 'GridSearchStrategy']
