+++ src/backend/logs/bayesian_search_DecisionTreeClassifier_20251019_001731.csv
@@ -1,0 +1,3 @@
+model,run_type,best_params,accuracy,precision_macro,precision_weighted,recall_macro,recall_weighted,f1_macro,f1_weighted
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 5, 'min_samples_split': 2}",0.8200000000000001,0.8301301476301477,0.8301301476301477,0.8200000000000001,0.8200000000000001,0.8188302345416358,0.8188302345416358
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 5, 'min_samples_split': 2}",0.8200000000000001,0.8301301476301477,0.8301301476301477,0.8200000000000001,0.8200000000000001,0.8188302345416358,0.8188302345416358
