+++ src/backend/logs/bayesian_search_DecisionTreeClassifier_20251019_002400.csv
@@ -1,0 +1,12 @@
+model,run_type,best_params,accuracy,precision_macro,precision_weighted,recall_macro,recall_weighted,f1_macro,f1_weighted
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 5, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 5, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 5, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 5, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
+DecisionTreeClassifier,bayesian_search,"{'max_depth': 3, 'min_samples_split': 2}",0.96,0.9666666666666668,0.9666666666666668,0.96,0.96,0.9595959595959596,0.9595959595959596
